+++ src/pages/LoginPage.tsx
@@ -1,5 +1,4 @@
-﻿import { useState, type FormEvent } from 'react'
-import { useNavigate, useLocation } from 'react-router-dom'
+﻿import { useNavigate, useLocation } from 'react-router-dom'
 import { useAuth } from '@/contexts/AuthContext'
 import styled from '@emotion/styled'
 import backIcon from '@/assets/back.png'
@@ -7,6 +6,7 @@ import loginIcon from '@/assets/user.png'
 import { colors } from '@/theme/color'
 import { typography } from '@/theme/typography'
 import { spacing } from '@/theme/spacing'
+import LoginFormSection from '@/components/LoginFormSection'
 
 const Container = styled.div`
   display: flex;
@@ -56,53 +56,14 @@ const Logo = styled.h2`
   color: ${colors.text.default};
 `
 
-const Form = styled.form`
-  width: 100%;
-  max-width: 360px;
-  display: flex;
-  flex-direction: column;
-`
-
-const Input = styled.input`
-  height: 48px;
-  padding: 0 ${spacing.spacing2};
-  margin-bottom: ${spacing.spacing4};
-  border: 1px solid ${colors.border.default};
-  border-radius: 4px;
-  font-size: ${typography.body1Regular.fontSize};
-  outline: none;
-
-  &::placeholder {
-    color: ${colors.text.placeholder};
-  }
-`
-
-const Button = styled.button`
-  height: 48px;
-  background-color: ${colors.brand.kakaoYellow};
-  border: none;
-  border-radius: 4px;
-  font-size: ${typography.body1Bold.fontSize};
-  font-weight: ${typography.body1Bold.fontWeight};
-  cursor: pointer;
-  transition: background 0.2s ease;
-
-  &:hover {
-    background-color: ${colors.brand.kakaoYellowHover};
-  }
-`
-
 export default function LoginPage() {
   const navigate = useNavigate()
   const location = useLocation()
   const { login } = useAuth()
   const from = (location.state as { from?: string })?.from ?? '/'
 
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
 
-  const handleSubmit = (e: FormEvent) => {
-    e.preventDefault()
+  const handleSuccess = () => {
     login()
     navigate(from, { replace: true })
   }
@@ -120,23 +81,7 @@ export default function LoginPage() {
       </Header>
       <Main>
         <Logo>kakao</Logo>
-        <Form onSubmit={handleSubmit}>
-          <Input
-            type="email"
-            placeholder="이메일"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-          <Input
-            type="password"
-            placeholder="비밀번호"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-          <Button type="submit">로그인</Button>
-        </Form>
+        <LoginFormSection onSuccess={handleSuccess} />
       </Main>
     </Container>
   )
